requirements.txt
keras_ocr==0.9.3
matplotlib==3.8.0
numpy==1.26.1
opencv_python==4.8.1.78
pandas==2.1.2
Pillow==9.0.1
presidio_image_redactor==0.0.49
pydicom==2.4.3
pytesseract==0.3.10
tensorflow==2.14.0
